refactor: checking for cyclic import
